requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
python-dotenv==1.2.1
langchain==1.2.10
langchain-groq==1.1.2
langchain-community==0.4.1
langgraph==1.0.8
chromadb==1.5.0
pydantic==2.12.5
sentence-transformers==5.2.2
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
